Move config to own package, add DialTimeout.
